Added a function to list all generals that are currently being used
Not sure if this really fits with the validator, or if it should be a separate thing, but one of the devs asked for it and I don't say no when a dev asks for something. And like, the easiest way seemed to be to make it a part of the validator so I could use the existing infrastructure, so here it is.